# app/static/swagger.yaml
openapi: "3.0.0"
info:
  title: "Seznam API"
  description: "This is documentation for Seznam API"
  version: "0.1"
  contact:
    email: "m.laskowski@o2.pl"
servers:
  - url: "/"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "All endpoints are authorized by access tokens.
      'Root' accepts the 'ROOT_PWD' token defined as an environmental variable.
      Rest of the endpoints accept tokens generated via 'root/create_token' endpoint."
security:
  - bearerAuth: [ ]
paths:
...
  /delete/file/:
    delete:
      tags:
        - "delete/file"
      summary: "Delete file"
      operationId: "delete/file/"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              required:
                - "path"
                - "filename"
              properties:
                path:
                  type: "string"
                  example: "dir1/dir2"
                filename:
                  type: "string"
                  example: "file.txt"
      responses:
        "200":
          description: "Data deleted properly"
        "400":
          description: "Request validation error"
        "401":
          description: "Token validation error"
        "404":
          description: "File not found"
  /delete/dir/:
    delete:
      tags:
        - "delete/dir"
      summary: "Delete empty directory"
      operationId: "delete/dir/"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              required:
                - "path"
                - "dirname"
              properties:
                path:
                  type: "string"
                  example: "dir1"
                dirname:
                  type: "string"
                  example: "dir2"
      responses:
        "200":
          description: "Data deleted properly"
        "400":
          description: "Request validation error or directory not empty"
        "401":
          description: "Token validation error"
        "404":
          description: "Directory not found"
...